pyproject.toml
[tool.poetry]
name = "rpi-webserver"
version = "0.1.0"
description = "Lightweight web server running on Raspberry Pi"
authors = ["Your Name <you@example.com>"]

packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.2"
flask-bcrypt = "^1.0"
flask-mongoengine-3 = "^1.1.0"
flask-jwt-extended = "^4.4.0"
Flask-Limiter = "^3.5"
bleach = "^6.1"
pydantic = "^2.0.0"
blinker = "^1.6.0"
password-strength = "^0.0.3"
gunicorn = "^21.2"
python-dotenv = "^1.1.1"
python-slugify = "^8.0.1" # Added for slugify functionality




[tool.poetry.dev-dependencies]
black = "^24.3"
flake8 = "^7.0"
pytest = "^8.0"
pytest-flask = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
